Add files via upload
Built an interactive resume in power bi.

Built the background template using figma. Created a summarized and detailed resume review with tooltips adding logos of institutions and software's to make the resume interactive.
Used bookmarks and tooltips to make the resume fun and interactive. Added a projects section to go with the digital resume for people to see at a glance my prominent projects